Changing file names and moving to another location. Adding implementation of methods for handling user-added data sets.
